move record sorting logic to deprecated namespace
